requirements.txt
asgiref==3.8.1
Django==5.2.1
django-environ==0.12.0
djangorestframework==3.16.0
psycopg2==2.9.10
sqlparse==0.5.3
tzdata==2025.2
